Vision-Z - kimi/requirements.txt
streamlit==1.28.2
opencv-python==4.8.1.78
ultralytics==8.0.200
torch==2.1.0
torchvision==0.16.0
psycopg2-binary==2.9.9
pyttsx3==2.90
python-dotenv==1.0.0
ollama==0.1.7
numpy==1.24.3
Pillow==10.0.1
